pyproject.toml
[build-system]
requires = ["setuptools>=62"]
build-backend = "setuptools.build_meta"

[project]
name = "autoapply"
version = "0.1.0"
description = "Automated job application agent with resume-aware prioritisation and browser automation"
authors = [{name = "Auto Generated"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pydantic>=1.10",
    "requests>=2.31",
    "python-dateutil>=2.8",
    "pypdf>=4.0",
    "Pillow>=10.0",
    "numpy>=1.24"
]

[project.scripts]
autoapply-agent = "autoapply.cli:main"

[project.optional-dependencies]
testing = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "responses>=0.25"
]
playwright = [
    "playwright>=1.43"
]

[tool.setuptools.packages.find]
where = ["src"]
